feat(backend): add auth middleware to Update Project and Delete Project routes

+++ server/src/router/projectRoutes.ts
@@ -39,6 +39,8 @@ router.post(
 // Update project
 router.patch(
   '/projects/:id',
+  authenticate,
+  authorize(GlobalRole.ADMIN),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await updateProject(req, res, next, prisma);
   }
@@ -47,6 +49,8 @@ router.patch(
 // Delete project
 router.delete(
   '/projects/:id',
+  authenticate,
+  authorize(GlobalRole.ADMIN),
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await deleteProject(req, res, next, prisma);
   }
